berghem_project/requirements.txt
django==2.2
djangorestframework==3.9.2
virtualenv==20.0.21
djongo==1.3
